+++ Assets/PlayFabEditorExtensions/Editor/Resources/MostRecentPackage.unitypackage.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 40c58c18d630d154ea5a29693a6f3a56
-NativeFormatImporter:
+guid: 86f5a61ef8ba75343aa7f91ddc6ad8f3
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
