+++ Assets/Scenes/ExampleScenes.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 40c58c18d630d154ea5a29693a6f3a56
-NativeFormatImporter:
+guid: e625fb11c2759a7468d4dc0aae99665f
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
