+++ Assets/Scenes/ExampleScenes/SceneLesson3.unity.meta
