+++ Assets/Scripts/ExampleScripts.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 40c58c18d630d154ea5a29693a6f3a56
-NativeFormatImporter:
+guid: 3aa33bb219d217941a5f905a60e2eb28
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
