+++ Assets/Scripts/ExampleScripts/PhotonLogin.cs
@@ -1,0 +1,37 @@
+using Photon.Pun;
+using UnityEngine;
+
+
+namespace ExampleGB
+{
+    public class PhotonLogin : MonoBehaviourPunCallbacks
+    {
+        private void Awake()
+        {
+            PhotonNetwork.AutomaticallySyncScene = true;
+        }
+
+        private void Start()
+        {
+            Connect();
+        }
+
+        public void Connect()
+        {
+            if (PhotonNetwork.IsConnected)
+            {
+                PhotonNetwork.JoinRandomRoom();
+            }
+            else
+            {
+                PhotonNetwork.ConnectUsingSettings();
+            }
+        }
+
+        public override void OnConnectedToMaster()
+        {
+            base.OnConnectedToMaster();
+            Debug.Log("Photon Success");
+        }
+    }
+}
