+++ Assets/Scripts/ExampleScripts/PhotonLogin.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: badc11d7192bf47c6a26d54e745bfc67
+guid: 6c1c133333a600b4d81e688b39769709
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
