+++ Assets/Scripts/ExampleScripts/PlayFabLogin.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 37de811dfb3cf4c2093720dd4215ab3a
+guid: e70ef3629d836d9429b683c7f9d2bccc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
