--- Assets/Scripts/PhotonLogin.cs
@@ -1,33 +1,0 @@
-using Photon.Pun;
-using UnityEngine;
-
-public class PhotonLogin : MonoBehaviourPunCallbacks
-{
-    private void Awake()
-    {
-        PhotonNetwork.AutomaticallySyncScene = true;
-    }
-
-    private void Start()
-    {
-        Connect();
-    }
-
-    public void Connect()
-    {
-        if (PhotonNetwork.IsConnected)
-        {
-            PhotonNetwork.JoinRandomRoom();
-        }
-        else
-        {
-            PhotonNetwork.ConnectUsingSettings();
-        }
-    }
-
-    public override void OnConnectedToMaster()
-    {
-        base.OnConnectedToMaster();
-        Debug.Log("Photon Success");
-    }
-}
